Supporting removal of translations
The internal structure of translations was a single object that was merged with any new data that was added. Since we want to allow users to add/remove translations in real time (and designer), changing the internal structure to be 2 objects: `translationSet` and `appLevelTranslations`.

`translationSet`: The translations that were received from the cloud configuration service for the network/site.

`appLevelTranslations`: Translations that were either set in the instance config or in designer. These will override the `translationSet` values.

Any adding/removal of translations happens in the `appLevelTranslations` tier and is always merged on top of the `translationSet` data. The resulting `translations` object is that current set of translations that are used in the SDK and apps.

#nojira